Separate ComputedModel::Model from ComputedModel
